Interim, working - HarmonicOscillator2 is working
Embedded control into the harmonic oscillator.  This is a breaking change, but definately cleans up the
boilerplate in anything intending to use harmonic oscillator. Using cycle-proxy to proxy the control stream. Tried
most-proxy but it isn't ready for prime time yet. Might switch to it when it is, as cycle-proxy uses several
adapter steps to be generic that probably slow things down.  Oscillator pauses whenever a rolling sum of the
absolute value of the speed is less than some threshold. It starts up again whenever a new governing funciton event
is provided.