Release 0.2.0 - Breaking change. Control is now embedded in HarmomicOscillator.
HarmonicOscillator auto-starts whenever there is a change in the input parameters, and auto-stops when
the rolling sum of the absolute value of the speed term is very small.  The former changes the governing
functions, creating a new dynamic problem to be solved, and the latter indicates that the solver has converged
on the solution.  HarmonicOscillator uses the midpoint method with h locked to the animation frame cycle.